Исправил проблему с запросом продуктов через API и сделал формы авторизации пользователя

+++ app/src/main/java/com/example/supporthealth/nutrition/search/data/dto/ProductDto.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class ProductDto(
     @SerializedName("product_id") val productId: String,
     val name: String,
-    val calories: Int,
+    val calories: Float,
     val protein: Float,
     val fat: Float,
     val carbs: Float
